app/src/main/java/com/ang/acb/addressbook/data/contacts/ContactsDataSource.kt
package com.ang.acb.addressbook.data.contacts

import kotlinx.coroutines.CoroutineDispatcher
import kotlinx.coroutines.Dispatchers
import kotlinx.coroutines.flow.Flow
import kotlinx.coroutines.withContext
import javax.inject.Inject

class ContactsDataSource @Inject constructor(
    private val contactDao: ContactDao,
    private val ioDispatcher: CoroutineDispatcher = Dispatchers.IO
) {

    suspend fun saveContact(
        firstName: String,
        lastName: String,
        email: String,
        phoneNumber: String,
        address: String,
    ): Long {
        return withContext(ioDispatcher) {
            contactDao.insert(
                ContactEntity(
                    firstName = firstName,
                    lastName = lastName,
                    email = email,
                    phoneNumber = phoneNumber,
                    address = address
                )
            )
        }
    }

    fun getContact(contactId: Long): Flow<ContactEntity?> {
        return contactDao.getContact(contactId)
    }

    fun getAllContacts(): Flow<List<ContactEntity>> {
        return contactDao.getAllContacts()
    }
}
